moved some custom polyglot functionality to its own util module
